requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
setuptools==69.1.1
sqlparse==0.4.4
wheel==0.43.0
